Final Project
Summary: There are many factors that lead to death for those with heart disease. Specifically we want to look at serum creatinine and ejection fraction. We want to identify what a normal range looks like for each, and see whether falling into the normal range or not can be used to correlate with death event. We also want to see whether serum creatinine or ejection fraction works better, as measured by correlation.
The implication of this work is that it could allow doctors to better predict the outcome of death based on test result from serum creatinine or ejection fraction. After our research, we came to the conclusion that both have negative correlations which gives us the logical conclusion that abnormal values lead to death events. However, both correlations are low, which tells as that this would not be an effective metric. However, previous results indicate that in tandem with other factors both serum creatinine and ejection fraction are useful, they just cannot be used individually.
As to which one is more effective, the efficacy of both are low but the correlation with normal serum creatinine and death event is better in comparison.